refactor: create a utility type to access table types
