--- apps/web/app/db/getPrismaClient.ts
@@ -1,9 +1,0 @@
-import { PrismaClient } from 'db'
-
-const globalForPrisma = globalThis as unknown as {
-  prisma: PrismaClient | undefined
-}
-
-export const prisma = globalForPrisma.prisma ?? new PrismaClient()
-
-if (process.env['NODE_ENV'] !== 'production') globalForPrisma.prisma = prisma
