+++ packages/db/src/domains/roles/getSubscriberRoles.ts
@@ -1,8 +1,9 @@
 import { Database, SupabaseClient } from 'db'
 import { Entities } from 'shared/src/enums/entities'
+import { SupabaseTable } from '../../supabase/utilityTypes'
 
 type EnglishLevel = Database['public']['Enums']['EnglishLevel']
-type Subscribers = Database['public']['Tables']['Subscribers']['Row']
+type Subscribers = SupabaseTable<'Subscribers'>
 
 const englishLevelScore: Record<EnglishLevel, number> = {
   Fluent: 3,
