+++ packages/db/src/supabase/type.ts
@@ -386,4 +386,3 @@ export interface Database {
     }
   }
 }
-
